LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/backdrop-filter-bounds-expansion-expected.txt


FAIL drawElementImage applies backdrop-filter bounds beyond child drawable bounds assert_array_equals: Backdrop filter bounds should be expanded to cover the whole element expected property 0 to be 0 but got 255 (expected array [0, 255, 255, 255] got object "255,0,0,255")
